refactor: split new feature switch into multiple files
