+++ frontend/src/component/project/Project/ProjectFeatureToggles/FeatureToggleSwitch/NewFeatureToggleSwitch.types.ts
@@ -1,0 +1,18 @@
+import { ReactNode } from 'react';
+
+export type OnFeatureToggleSwitchArgs = {
+    featureId: string;
+    projectId: string;
+    environmentName: string;
+    environmentType?: string;
+    hasStrategies?: boolean;
+    hasEnabledStrategies?: boolean;
+    isChangeRequestEnabled?: boolean;
+    onRollback?: () => void;
+    onSuccess?: () => void;
+};
+
+export type UseFeatureToggleSwitchType = (projectId: string) => {
+    modals: ReactNode;
+    onToggle: (newState: boolean, config: OnFeatureToggleSwitchArgs) => void;
+};
